Make tester knob protection private
